+++ website/docs/components/accordion/examples/one_section_opening.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import Accordion from 'intergalactic/accordion';
-import { Text } from 'intergalactic/typography';
-import { Box, Flex } from 'intergalactic/flex-box';
+import { Box } from 'intergalactic/flex-box';
 
 const Demo = () => {
   const [value, onChange] = React.useState(null); // or []
@@ -10,10 +9,10 @@ const Demo = () => {
       {[...new Array(3)].map((_, index) => (
         <Accordion.Item value={index} key={index} disabled={index === 2}>
           <Accordion.Item.Toggle pb={2}>
-            <Flex alignItems='center'>
+            <Accordion.Item.ToggleButton tag={'h3'}>
               <Accordion.Item.Chevron mr={2} />
-              <Text size={300} tag='h3' my={0}>{`Section ${index + 1}`}</Text>
-            </Flex>
+              Section {index + 1}
+            </Accordion.Item.ToggleButton>
           </Accordion.Item.Toggle>
           <Accordion.Item.Collapse>
             <Box p='12px 24px 24px'>{`Hello Section ${index + 1}`}</Box>