+++ semcore/select/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.45.4] - 2024-08-13
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/dropdown-menu` [4.35.3 ~> 4.35.4]).
+
+## [4.45.3] - 2024-08-12
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/dropdown-menu` [4.35.2 ~> 4.35.3], `@semcore/popper` [5.39.2 ~> 5.39.3]).
+
 ## [4.45.2] - 2024-07-31
 
 ### Fixed